requirements.txt
asgiref
channels
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
cffi==1.11.5
Django==2.1.5
django-celery-beat==1.4.0
django-timezone-field==3.0
djangorestframework==3.4.6
djangorestframework-expiring-authtoken==0.1.4
gevent==1.4.0
greenlet==0.4.15
kombu==4.2.2.post1
mysqlclient==1.3.14
pycparser==2.19
python-crontab==2.3.6
python-dateutil==2.7.5
pytz==2018.9
six==1.12.0
vine==1.2.0
